html/dom/elements/global-attributes/the-lang-attribute-003.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8"/>
<title>HTTP header</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='https://html.spec.whatwg.org/multipage/dom.html#the-lang-and-xml:lang-attributes'>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<link rel="stylesheet" type="text/css" href="../../../../testharness.css">
<meta name="assert" content="The browser will recognize a language declared in the HTTP header, when there is no internal language declaration.">
<meta name='flags' content='http dom :lang'>
<style type='text/css'>
#colonlangcontroltest { color: red; font-weight: bold; width: 400px; }
#colonlangcontroltest:lang(xx) { display:none; }
.test div { width: 50px; }
#box:lang(ko) { width: 100px; }
</style>
</head>
<body>
<div class="test"><div id="box">&#xA0;</div></div>
<p lang='xx' id='colonlangcontroltest'>This test failed because it relies on :lang for results, but :lang is not supported by this browser.</p>
<!--Notes:
This test uses :lang to detect whether the language has been set. If :lang is not supported, a message will appear and the test will fail.
The HTTP headers include the following: Content-Language: ko
Spec text: «If neither the node nor any of the node's ancestors, including the root element, have either attribute set, but there is a pragma-set default language set, then that is the language of the node. If there is no pragma-set default language set, then language information from a higher-level protocol (such as HTTP), if any, must be used as the final fallback language instead.»
-->
<script>
test(function() {
assert_equals(document.getElementById('colonlangcontroltest').offsetWidth, 0)
assert_equals(document.getElementById('box').offsetWidth, 100);
}, " ");
</script>
<div id='log'></div>
</body>
</html>
