third_party/WebKit/LayoutTests/imported/web-platform-tests/html/dom/elements/global-attributes/the-lang-attribute-003-expected.txt

This is a testharness.js-based test.
FAIL The browser will recognize a language declared in the HTTP header, when there is no internal language declaration. assert_equals: expected 100 but got 50
Harness: the test ran to completion.
